add: common useMutation api
